% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aBCF.R
\name{aBCF}
\alias{aBCF}
\title{Fit a single chain of an aBCF model}
\usage{
aBCF(
  Y_train,
  treated,
  obs_weights,
  X_cont_mu = matrix(0, nrow = 1, ncol = 1),
  X_cat_mu = matrix(0, nrow = 1, ncol = 1),
  X_cont_tau = matrix(0, nrow = 1, ncol = 1),
  X_cat_tau = matrix(0, nrow = 1, ncol = 1),
  unif_cuts_mu = rep(TRUE, times = ncol(X_cont_mu)),
  unif_cuts_tau = rep(TRUE, times = ncol(X_cont_tau)),
  cutpoints_list_mu = NULL,
  cutpoints_list_tau = NULL,
  cat_levels_list_mu = NULL,
  cat_levels_list_tau = NULL,
  sparse = FALSE,
  M_mu = 200,
  M_tau = 50,
  sd_mu = 2,
  alpha_mu = 0.95,
  beta_mu = 2,
  sd_tau = 1,
  alpha_tau = 0.25,
  beta_tau = 3,
  min_node_size = 5,
  use_halfnormal_scales = TRUE,
  sigu_hyperprior = 2/3,
  nd = 1000,
  burn = 1000,
  thin = 1,
  save_samples = TRUE,
  batch_size = 100,
  acceptance_target = 0.44,
  nu = 3,
  lambda = NULL,
  prior_only = FALSE,
  verbose = TRUE,
  print_every = floor((nd * thin + burn))/10,
  perm = NULL,
  chain_num = 1
)
}
\arguments{
\item{Y_train}{Vector of outcome values; will be standardized internally.}

\item{treated}{Vector of indicators for treated units}

\item{obs_weights}{Vector of weights, used for calculating heteroskedatic errors}

\item{X_cont_mu}{Matrix of continuous predictors for use in fitting mu trees; must be scaled to [-1,1]}

\item{X_cat_mu}{Matrix of categorical predictors for mu trees}

\item{X_cont_tau}{Matrix of continuous predictors for tau trees; must be scaled to [-1,1]}

\item{X_cat_tau}{Matrix of categorical predictors for tau trees}

\item{unif_cuts_mu}{Logical vector of length ncol(X_cont_mu) for wheter or not to use uniform cutpoints for continuous mu predictors}

\item{unif_cuts_tau}{Logical vector of length ncol(X_cont_tau) for wheter or not to use uniform cutpoints for continuous tau predictors}

\item{cutpoints_list_mu}{List of cutpoints for continuous mu predictors}

\item{cutpoints_list_tau}{List of cutpoints for continuous tau predictors}

\item{cat_levels_list_mu}{List of levels for each categorical mu predictor}

\item{cat_levels_list_tau}{List of levels for each categorical tau predictor}

\item{sparse}{Logical; if TRUE uses Dirichlet variable selection prior, otherwise uses BART's uniform random variable selection}

\item{M_mu}{Number of mu trees}

\item{M_tau}{Number of tau trees}

\item{sd_mu}{Prior scale of the leaf node means of mu trees}

\item{alpha_mu}{Base for tree prior on mu trees}

\item{beta_mu}{Power for tree prior on mu trees}

\item{sd_tau}{Prior scale of the leaf node means of tau trees}

\item{alpha_tau}{Base for tree prior on tau trees}

\item{beta_tau}{Power for tree prior on tau trees}

\item{min_node_size}{Minimum sample size in each end node.}

\item{use_halfnormal_scales}{Logical; whether or not to use halfnormal scale priors for mu and tau SDs}

\item{sigu_hyperprior}{Standard deviation of half-normal hyperprior for sigma_u, given in units of sd(y)}

\item{nd}{Number of posterior samples to return}

\item{burn}{Number of burn-in iterations}

\item{thin}{Thinning factor for posterior; burn + nd*thin total iterations are performed.}

\item{save_samples}{Whether or not to return results. No clue why you'd ever set to FALSE}

\item{batch_size}{Batch size for calculating adaptive MH acceptance of sigmas}

\item{acceptance_target}{Targeted acceptance rate for adaptive MH}

\item{nu}{Degrees of freedom in the chisq prior on \eqn{sigma^2}}

\item{lambda}{Scale parameter in the chisq prior on \eqn{sigma^2}}

\item{prior_only}{Logical; whether to actually fit model or just draw from prior}

\item{verbose}{Whether to print output to console}

\item{print_every}{Print to console every print_every'th iteration}

\item{perm}{Permutation to use to sort data}

\item{chain_num}{Chain identifier; only used if called in multi-chain run}
}
\description{
Fit a single chain of an aBCF model
}
