% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_fits.R
\name{get_tree_fits}
\alias{get_tree_fits}
\title{Generate predictions from a fitted flexBCFa model}
\usage{
get_tree_fits(
  fit,
  type = c("mu", "tau"),
  X_cont = matrix(0, nrow = 1, ncol = 1),
  X_cat = matrix(0, nrow = 1, ncol = 1),
  verbose = TRUE,
  print_every = floor(nrow(fit$mu)/10),
  n_cores = 1
)
}
\arguments{
\item{fit}{The fitted model object}

\item{type}{Character vector of which type of trees (mu, tau) to generate prediction for.}

\item{X_cont}{Matrix of continuous predictors to use. If you want to use different predictors
for mu vs tau you will need to run predict separately
NB: predictors must be scaled to [-1,1]; this must be done using the same scaling as was used for fitting the original model}

\item{X_cat}{Matrix of categorical predictors}

\item{verbose}{Whether to print verbosely}

\item{print_every}{Print to log after each print_every'th iteration}

\item{n_cores}{Number of cores to use for prediction. Should be <= number of chains}
}
\description{
Generate predictions from a fitted flexBCFa model
}
